Break algorithms into separate files.  With tests on each digest.
